+++ src/widgets/ColorButton.cpp
@@ -1,0 +1,53 @@
+/*
+ * Copyright (C) 2017 Damir Porobic <https://github.com/damirporobic>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301, USA.
+ */
+
+#include "ColorButton.h"
+
+ColorButton::ColorButton(QWidget* parent) : QPushButton(parent)
+{
+    connect(this, &ColorButton::clicked, this, &ColorButton::openDialog);
+}
+
+QColor ColorButton::color() const
+{
+    return mColor;
+}
+
+void ColorButton::setColor(QColor&& color)
+{
+    mColor = color;
+    auto pixmapIcon = createPixmapFromColor(qMove(mColor));
+    setIcon(pixmapIcon);
+}
+
+QPixmap ColorButton::createPixmapFromColor(QColor&& color)
+{
+    QPixmap pixmap(32, 32);
+    pixmap.fill(color);
+    return pixmap;
+}
+
+void ColorButton::openDialog()
+{
+    auto color = QColorDialog::getColor(mColor, parentWidget());
+    if (!color.isValid()) {
+        return;
+    }
+    setColor(qMove(color));
+}